requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.1
django-admin-bootstrapped==2.5.0
django-bootstrap3==5.4.0
django-crispy-forms==1.4.0
django-haystack==2.3.1
django-toolbelt==0.0.1
django-widget-tweaks==1.3
gunicorn==19.3.0
linecache2==1.0.0
Pillow==2.8.2
psycopg2==2.6
python-gettext==2.1
six==1.9.0
static3==0.6.0
traceback2==1.4.0
unittest2==1.0.1
whitenoise==1.0.6
Whoosh==2.7.0
